requirements.txt

#    This file is part of OnDA.
#
#    OnDA is free software: you can redistribute it and/or modify
#    it under the terms of the GNU General Public License as published by
#    the Free Software Foundation, either version 3 of the License, or
#    (at your option) any later version.
#
#    OnDA is distributed in the hope that it will be useful,
#    but WITHOUT ANY WARRANTY; without even the implied warranty of
#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
#    GNU General Public License for more details.
#
#    You should have received a copy of the GNU General Public License
#    along with OnDA.  If not, see <http://www.gnu.org/licenses/>.
#
#    Copyright 2014-2018 Deutsches Elektronen-Synchrotron DESY,
#    a research centre of the Helmholtz Association.cfelpyutils>=0.9
fabio[monitor-cbf]>=0.6.0
h5py>=2.7.0
mpi4py>=2.0.0
numpy>=1.11.3
pyqt[gui-qt4]>=4.11.4
pyqt[gui-qt5]>=5.9.2
pyqtgraph[gui-qt4]>=0.10.0
pyqtgraph[gui-qt5]>=0.10.0
psana-conda[monitor-psana]>=1.3.54
scipy>=1.1.0
toml>=0.9.6
msgpack-python>=0.5.6
msgpack-numpy>=0.4.3.1

